Fix save sync on systems without libadwaita
One of the methods used for this contained an import statement for
`gi.repository.Awd`, which fails if libadwaita is not available. This
import is also already done at the top level, and `Adw` is mocked there
if it's not available, so simply removing this import is enough to get
this to work.